requirements.txt
boto3
python-dotenv
requests
requests-cache
tenacity
schedule
sqlalchemy